requirements.txt
Django==1.9.5
MySQL-python==1.2.5
Pillow==3.2.0
Unidecode==0.04.19
Willow==0.3
argparse==1.2.1
beautifulsoup4==4.4.1
boto==2.38.0
django-appconf==1.0.1
django-compressor==2.0
django-ckeditor==5.0.3
django-debug-toolbar==1.4
django-forms-bootstrap==3.0.1
django-modelcluster==1.1
django-overextends==0.4.1
django-redis-cache==1.6.5
django-storages==1.1.8
django-taggit==0.18.1
django-treebeard==4.0
djangorestframework==3.3.3
elasticsearch==2.3.0
html5lib==1.0b8
opbeat==3.3.1
pytz==2016.3
rcssmin==1.0.6
redis==2.10.5
rjsmin==1.0.12
setproctitle==1.1.9
six==1.10.0
sphinx-me==0.3
urllib3==1.14
wagtail==1.5.2
wagtailsettings==0.4.1
wsgiref==0.1.2
raven